refactor: rename vscode -> vscode_nvim
