+++ config/nvim/lua/core/keymap.lua
@@ -6,3 +6,7 @@ vim.api.nvim_set_keymap('i', '<C-o>', '<Esc>o', { noremap = true, silent = true 
 
 -- remove highlight
 vim.api.nvim_set_keymap('n', '<leader>h', ':noh<cr>', { noremap = true, silent = true })
+
+-- move
+vim.api.nvim_set_keymap('n', 'H', '0', { noremap = true, silent = true })
+vim.api.nvim_set_keymap('n', 'L', '$', { noremap = true, silent = true })
